move restore volume api tests to separate folder
api test file was getting too big
